requirements.txt
transformers==4.15.0
torch==1.10.1
construct==2.8.14
mne==1.0.2
PsychoPy==2021.2.3
openpyxl==2.6.3
numpy==1.20.3
sounddevice==0.4.1
SoundFile==0.10.3.post1
scipy==1.5.2
scikit-learn==0.23.2
seaborn==0.9.0
matplotlib==3.4.2
pylsl==1.14.0
pandas==1.1.3
psutil==5.7.2
Pillow==9.0.1
py-cpuinfo==7.0.0
pyedflib==0.1.30
PyQt5==5.15.1
tqdm==4.62.2
